REFACTOR: Moving events components over to the events folder
