+++ app/api/events/featured/route.ts
