+++ app/events/_components/EventClient.tsx
