+++ app/events/_components/SearchBar.tsx
