+++ app/page.tsx
@@ -1,7 +1,3 @@
-import HeroSection from "./_components/home/Hero";
-import FeaturedEventsSection from "./_components/home/FeaturedEvents";
-import AboutSection from "./_components/home/About";
-import SponsorsSection from "./_components/home/Sponsors";
 import HomeClient from "./_components/home/HomeClient";
 
 export default async function Home() {
@@ -15,15 +11,4 @@ export default async function Home() {
     const events = json.data;
 
     return <HomeClient events={events} />;
-    // (
-    //     <>
-    //         <HeroSection />
-    //         <FeaturedEventsSection events={events} />
-    //         <div className="divider"></div>
-    //         <AboutSection />
-    //         <div className="divider"></div>
-    //         <SponsorsSection />
-    //     </>
-        
-    // );
 }
